remove artifacts from old structure, add new file
